+++ src/main/java/com/CuteNekoDragon/Core/config/ClientConfig.java
@@ -1,0 +1,8 @@
+package com.CuteNekoDragon.Core.config;
+
+import net.minecraftforge.common.ForgeConfigSpec;
+
+public final class ClientConfig {
+
+    ClientConfig(ForgeConfigSpec.Builder builder) {}
+}