+++ src/main/java/com/CuteNekoDragon/Core/config/CommonConfig.java
@@ -1,0 +1,8 @@
+package com.CuteNekoDragon.Core.config;
+
+import net.minecraftforge.common.ForgeConfigSpec;
+
+public final class CommonConfig {
+
+    CommonConfig(ForgeConfigSpec.Builder builder) {}
+}