+++ src/main/java/com/CuteNekoDragon/Core/config/SVOConfig.java
@@ -1,9 +1,10 @@
-package com.CuteNekoDragon.svo_core.config;
+package com.CuteNekoDragon.Core.config;
 
 import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.config.ModConfigEvent;
+
 import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.function.Function;
@@ -14,14 +15,10 @@ public final class SVOConfig {
     public static final CommonConfig COMMON = register(ModConfig.Type.COMMON, CommonConfig::new);
     public static final ServerConfig SERVER = register(ModConfig.Type.SERVER, ServerConfig::new);
 
-    public static void init() {
-
-    }
+    public static void init() {}
 
     @SuppressWarnings("unused")
-    public static void onLoad(final ModConfigEvent event) {
-
-    }
+    public static void onLoad(final ModConfigEvent event) {}
 
     @SuppressWarnings("removal")
     public static <C> C register(ModConfig.Type type, Function<ForgeConfigSpec.Builder, C> factory) {